Defer breaking up `auth.ts`
